# src/main/resources/api.yml
openapi: 3.0.1
info:
  title: Demo
  description: 'Demo'
  version: 0.5.0
servers:
  - url: https://op.op.com
...
paths:
...
  /reporte:
    get:
      tags:
        - Reporte
      summary: Retorna todos los reportes
      operationId: getAllReportes
      responses:
        200:
          description: Operacion Exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportesResponse'
        401:
          description: No Autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        500:
          description: Error en el servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
    post:
      tags:
        - Reporte
      summary: Crea un reporte
      operationId: saveReporte
      requestBody:
        description: Crea un nuevo reporte
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReporteRequest'
        required: true
      responses:
        200:
          description: Reporte creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReporteResponse'
        401:
          description: No Autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        500:
          description: Error en el servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /reporte/telefono/{telefono}:
    get:
      tags:
        - Reporte
      summary: Retorna los reportes de un ciudadano especifco
      operationId: getReportesByTelefono
      parameters:
        - name: telefono
          in: path
          description: telefono del ciudadano
          required: true
          schema:
            type: string
      responses:
        200:
          description: Operacion Exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportesResponse'
        401:
          description: No Autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        500:
          description: Error en el servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /reporte/{id}:
    get:
      tags:
        - Reporte
      summary: Retorna un Reporte especifico
      operationId: getReporteById
      parameters:
        - name: id
          in: path
          description: ID del reporte a retornar
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Reporte encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReporteResponse'
        401:
          description: No Autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        404:
          description: No se encontro el reporte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        500:
          description: Error en el servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
    put:
      tags:
        - Reporte
      summary: Actualizar un Reporte
      operationId: updateReporte
      parameters:
        - name: id
          in: path
          description: ID del reporte a editar
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Actualizar un reporte existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReporteRequest'
        required: true
      responses:
        200:
          description: Reporte actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        401:
          description: No Autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        404:
          description: No se encontro el reporte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        500:
          description: Error en el servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /reporte/location:
    post:
      tags:
        - Reporte
      summary: Retorna si es posible realizar un reporte en una ubicacion
      operationId: getPuedeCrearReporteEnUbicacion
      requestBody:
        description: Ubicacion a validar
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationRequest'
        required: true
      responses:
        200:
          description: Es Posible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UbicacionReporteResponse'
        400:
          description: Request Invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        401:
          description: No Autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        500:
          description: Error en el servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
...
  /tecnico:
    get:
      tags:
        - Technician
      summary: Retorna todos los tecnicos
      operationId: getAllTechnicians
      responses:
        200:
          description: Operacion Exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TechniciansResponse'
        401:
          description: No Autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        500:
          description: Error en el servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
    post:
      tags:
        - Technician
      summary: Crea un tecnico
      operationId: saveTechnician
      requestBody:
        description: Crea un nuevo tecnico
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TechnicianRequest'
        required: true
      responses:
        200:
          description: Technician creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TechnicianResponse'
        401:
          description: No Autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        500:
          description: Error en el servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /tecnico/telefono/{telefono}:
    get:
      tags:
        - Technician
      summary: Retorna un Technician especifico
      operationId: getTechnicianByTelefono
      parameters:
        - name: telefono
          in: path
          description: Telefono del tecnico a retornar
          required: true
          schema:
            type: string
      responses:
        200:
          description: Technician encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TechnicianResponse'
        401:
          description: No Autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        404:
          description: No se encontro el reporte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        500:
          description: Error en el servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /tecnico/{id}:
    get:
      tags:
        - Technician
      summary: Retorna un Technician especifico
      operationId: getTechnicianById
      parameters:
        - name: id
          in: path
          description: ID del tecnico a retornar
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Technician encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TechnicianResponse'
        401:
          description: No Autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        404:
          description: No se encontro el reporte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        500:
          description: Error en el servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
    put:
      tags:
        - Technician
      summary: Actualizar un Technician
      operationId: updateTechnician
      parameters:
        - name: id
          in: path
          description: ID del tecnico a editar
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Actualizar un tecnico existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TechnicianRequest'
        required: true
      responses:
        200:
          description: Asistencia actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        401:
          description: No Autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        404:
          description: No se encontro el reporte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        500:
          description: Error en el servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /tecnico/{id}/location:
    patch:
      tags:
        - Technician
      summary: Actualizar ubicacion del Technician
      operationId: updateTechnicianLocation
      parameters:
        - name: id
          in: path
          description: ID del tecnico a actualizar
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Actualizar un tecnico existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationRequest'
        required: true
      responses:
        200:
          description: Asistencia actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        401:
          description: No Autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        404:
          description: No se encontro el reporte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        500:
          description: Error en el servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
...
components:
  schemas:
...
    ReporteRequest:
      type: object
      required:
        - descripcion
        - longitud
        - latitud
        - telefono
      properties:
        telefono:
          type: string
        descripcion:
          type: string
        tipoReporte:
          type: string
          enum:
            - Hoyo en Pavimento
            - Otro
        longitud:
          type: string
        latitud:
          type: string
        notas:
          type: string
        fotos:
          type: array
          items:
            type: string
    Reporte:
      type: object
      properties:
        id:
          type: integer
          format: int64
        descripcion:
          type: string
        tipoReporte:
          type: string
        longitud:
          type: string
        latitud:
          type: string
        notas:
          type: string
        nombre:
          type: string
        apellido:
          type: string
        correo:
          type: string
        telefono:
          type: string
        fechaCreacion:
          type: string
          format: date-time
        fechaUltActualizacion:
          type: string
          format: date-time
        fotos:
          type: array
          items:
            type: string
        estado:
          $ref: '#/components/schemas/EstadoReporte'
    ReporteResponse:
      type: object
      properties:
        value:
          $ref: '#/components/schemas/Reporte'
        type:
          type: string
        message:
          type: string
    ReportesResponse:
      type: object
      properties:
        value:
          type: array
          items:
            $ref: '#/components/schemas/Reporte'
        type:
          type: string
        message:
          type: string
...
    TechnicianRequest:
      type: object
      required:
        - nombre
        - apellido
        - telefono
        - latitud
        - longitud
      properties:
        nombre:
          type: string
          maxLength: 100
        apellido:
          type: string
          maxLength: 100
        telefono:
          type: string
          maxLength: 100
        latitud:
          type: string
          maxLength: 50
        longitud:
          type: string
          maxLength: 50
    Technician:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
          maxLength: 100
        apellido:
          type: string
          maxLength: 100
        telefono:
          type: string
          maxLength: 100
        latitud:
          type: string
          maxLength: 50
        longitud:
          type: string
          maxLength: 50
    TechnicianResponse:
      type: object
      properties:
        value:
          $ref: '#/components/schemas/Technician'
        type:
          type: string
        message:
          type: string
    TechniciansResponse:
      type: object
      properties:
        value:
          type: array
          items:
            $ref: '#/components/schemas/Technician'
        type:
          type: string
        message:
          type: string
    UbicacionReporteResponse:
      type: object
      properties:
        value:
          $ref: '#/components/schemas/UbicacionReporte'
        message:
          type: string
    UbicacionReporte:
      type: object
      properties:
        posible:
          type: boolean
        message:
          type: string
    LocationRequest:
      type: object
      required:
        - latitud
        - longitud
      properties:
        latitud:
          type: string
          maxLength: 50
        longitud:
          type: string
          maxLength: 50
    EstadoReporte:
      type: object
      properties:
        descripcion:
          type: string
          maxLength: 50
        descripcionSecundaria:
          type: string
          maxLength: 50
        porcentajeCompletado:
          type: number
          format: float
        color:
          type: string
        detalle:
          type: array
          items:
            $ref: '#/components/schemas/DetalleEstadoReporte'
    DetalleEstadoReporte:
      type: object
      properties:
        nombre:
          type: string
          maxLength: 50
        fechaInicio:
          type: string
          format: date-time
        fechaFinal:
          type: string
          format: date-time
...
    EmptyResponse:
      type: object
      properties:
        type:
          type: string
        message:
          type: string